- remapped pins to my own board layout - fixed some bugs - support for pro mini board added

+++ OpenSourceLights/aa_light_setup.h
@@ -1,17 +1,19 @@
 #include <Arduino.h>
 #include <avr/pgmspace.h>
 
-extern const uint16_t NA;
-extern const uint16_t BLINK;
-extern const uint16_t FASTBLINK;
-extern const uint16_t SOFTBLINK;
-extern const uint16_t DIM;
-extern const uint16_t FADEOFF;
-extern const uint16_t XENON;
-extern const uint16_t BACKFIRE;
+extern const int NA;
+extern const int BLINK;
+extern const int FASTBLINK;
+extern const int SOFTBLINK;
+extern const int DIM;
+extern const int FADEOFF;
+extern const int XENON;
+extern const int BACKFIRE;
+
 extern const byte ON;
 extern const byte OFF;
+
 extern const byte YES;
 extern const byte NO;
 
-extern const uint16_t Schemes[NumSchemes][NumLights][NumStates] PROGMEM;
+extern const int Schemes[NumSchemes][NumLights][NumStates] PROGMEM;
